feat Lotto : 수익률을 소숫점 첫번째 자리까지 출력하도록 String.format()추가

+++ src/main/java/lotto/ProfitCalc.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import static java.lang.Math.round;
+
 public class ProfitCalc {
     float profit=0;
     public float profitCalc(int won){
@@ -7,7 +9,8 @@ public class ProfitCalc {
         int prize=0;
         int price=won;
 
-        profit=((prize-price)/price)*100;
+        profit=(float)(prize - price)/price * 100;
+
         return profit;
     }
 }
